separated the calculation of the Collateral Composition into an independent hook that returns an array of BigIntegers
